Updated src folder
The src folder has been updated to reflect the current version
of the code that is ready for use.

+++ src/initialise.py
@@ -11,7 +11,8 @@ def run():
     log.info("Peakfinder initialised.\n")
     
     # Removes previously created folders.
-    if os.path.exists("data"):
-        shutil.rmtree("data")
+    if ip.run_soh is True:
+        if os.path.exists("data"):
+            shutil.rmtree("data")
 
     return
